Move issue list queries to under the `issue/list` package
